# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  sourceUrl: string;
  industry: string;
  audience: string;
  tone: string;
  usp: string;
  values?: string[];
  status: string;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sourceUrl: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  industry?: string;
  audience?: string;
  tone?: string;
  usp?: string;
  values?: string[];
  status?: string;
}

export interface AnalysisInput {
  /** @minLength 1 */
  sourceUrl: string;
}

export interface Post {
  id: number;
  projectId: number;
  title: string;
  goal: string;
  format: string;
  text: string;
  /** @nullable */
  imageUrl?: string | null;
  status: string;
  variant: number;
  createdAt: string;
}

export interface PostInput {
  projectId: number;
  goal: string;
  format: string;
  /** @minLength 1 */
  topic: string;
}

export interface PostUpdate {
  title?: string;
  text?: string;
  status?: string;
}

export interface CompetitorPost {
  id: number;
  text: string;
  postedAt: string;
  idea: string;
}

export interface Competitor {
  id: number;
  projectId: number;
  name: string;
  url: string;
  lastSync: string;
  posts: CompetitorPost[];
}

export interface CompetitorInput {
  projectId: number;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  url: string;
}

export interface SyncInput {
  projectId: number;
}

export interface ScheduleItem {
  id: number;
  projectId: number;
  postId: number;
  /** @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$ */
  date: string;
  time: string;
  status: string;
  post: Post;
}

export interface ScheduleInput {
  projectId: number;
  postId: number;
  /** @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$ */
  date: string;
  time: string;
  channelIds?: number[];
}

export interface ScheduleUpdate {
  /** @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$ */
  date?: string;
  time?: string;
  status?: string;
}

export interface Dashboard {
  project: Project;
  totalPosts: number;
  approvedPosts: number;
  scheduledPosts: number;
  radarIdeas: number;
  recentPosts: Post[];
}

export type SocialChannelPlatform = typeof SocialChannelPlatform[keyof typeof SocialChannelPlatform];


export const SocialChannelPlatform = {
  vk: 'vk',
  ok: 'ok',
  max: 'max',
  telegram: 'telegram',
  zen: 'zen',
} as const;

export interface SocialChannel {
  id: number;
  platform: SocialChannelPlatform;
  name: string;
  target: string;
  status: string;
  /** @nullable */
  lastError?: string | null;
  createdAt: string;
}

export type SocialChannelInputPlatform = typeof SocialChannelInputPlatform[keyof typeof SocialChannelInputPlatform];


export const SocialChannelInputPlatform = {
  vk: 'vk',
  ok: 'ok',
  max: 'max',
  telegram: 'telegram',
  zen: 'zen',
} as const;

export interface SocialChannelInput {
  platform: SocialChannelInputPlatform;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  target: string;
  /** @minLength 1 */
  token: string;
  applicationKey?: string;
  applicationSecret?: string;
  sessionSecret?: string;
}

export interface PublishingJob {
  id: number;
  postId: number;
  channelId: number;
  scheduledAt: string;
  status: string;
  /** @nullable */
  externalId?: string | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  publishedAt?: string | null;
}

export interface PublishInput {
  postId: number;
  /** @minItems 1 */
  channelIds: number[];
}

export type PublishResultResultsItem = {
  channelId: number;
  status: string;
  /** @nullable */
  externalId?: string | null;
  error?: string;
};

export interface PublishResult {
  results: PublishResultResultsItem[];
}

export type ListPostsParams = {
/**
 * @nullable
 */
projectId?: number | null;
/**
 * @nullable
 */
status?: string | null;
};

export type ListCompetitorsParams = {
projectId: number;
};

export type ListScheduleParams = {
projectId: number;
};

export type TestChannel200 = {
  ok: boolean;
  message: string;
};