# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels:
    get:
      operationId: listChannels
      tags: [publishing]
      summary: List connected social channels
      responses:
        "200":
          description: Connected channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialChannel"
    post:
      operationId: createChannel
      tags: [publishing]
      summary: Connect a social channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialChannelInput"
      responses:
        "201":
          description: Connected channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialChannel"
  /channels/{id}:
    delete:
      operationId: deleteChannel
      tags: [publishing]
      summary: Disconnect a social channel
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Disconnected
  /channels/{id}/test:
    post:
      operationId: testChannel
      tags: [publishing]
      summary: Check a social channel connection
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Connection check
          content:
            application/json:
              schema:
                type: object
                required: [ok, message]
                properties:
                  ok: { type: boolean }
                  message: { type: string }
...
  /publishing/publish:
    post:
      operationId: publishPost
      tags: [publishing]
      summary: Publish a post to selected channels now
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishInput"
      responses:
        "200":
          description: Publication results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    SocialChannel:
      type: object
      required: [id, platform, name, target, status, createdAt]
      properties:
        id: { type: integer }
        platform:
          type: string
          enum: [vk, ok, max, telegram, zen]
        name: { type: string }
        target: { type: string }
        status: { type: string }
        lastError: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
    SocialChannelInput:
      type: object
      required: [platform, name, target, token]
      properties:
        platform:
          type: string
          enum: [vk, ok, max, telegram, zen]
        name: { type: string, minLength: 1 }
        target: { type: string, minLength: 1 }
        token: { type: string, minLength: 1 }
        applicationKey: { type: string }
        applicationSecret: { type: string }
        sessionSecret: { type: string }
...
    PublishInput:
      type: object
      required: [postId, channelIds]
      properties:
        postId: { type: integer }
        channelIds:
          type: array
          minItems: 1
          items: { type: integer }
    PublishResult:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            type: object
            required: [channelId, status]
            properties:
              channelId: { type: integer }
              status: { type: string }
              externalId: { type: [string, "null"] }
              error: { type: string }