# lib/api-zod/src/generated/types/publishResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublishResultResultsItem } from './publishResultResultsItem';

export interface PublishResult {
  results: PublishResultResultsItem[];
}