# lib/api-zod/src/generated/types/socialChannel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialChannelPlatform } from './socialChannelPlatform';

export interface SocialChannel {
  id: number;
  platform: SocialChannelPlatform;
  name: string;
  target: string;
  status: string;
  /** @nullable */
  lastError?: string | null;
  createdAt: Date;
}