# lib/api-zod/src/generated/types/socialChannelInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialChannelInputPlatform } from './socialChannelInputPlatform';

export interface SocialChannelInput {
  platform: SocialChannelInputPlatform;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  target: string;
  /** @minLength 1 */
  token: string;
  applicationKey?: string;
  applicationSecret?: string;
  sessionSecret?: string;
}